# openapi.yaml
openapi: 3.1.0
info:
  title: Pulse Payment Platform API
  description: |
    REST API for the Pulse Payment Platform. Create payment links to accept
    crypto (USDC/USDT on Solana) and PIX payments. Manage webhook subscriptions
    for real-time payment notifications.

    ## Authentication
    All endpoints require an API key sent as a Bearer token:
    ```
    Authorization: Bearer sk_live_...
    ```
    API keys are created in the Pulse dashboard under **Developers > API Keys**.

    ## Rate Limiting
    Responses include rate limit headers:
    - `X-RateLimit-Limit` — max requests per minute for your tier
    - `X-RateLimit-Remaining` — requests remaining in current window
    - `X-RateLimit-Reset` — Unix timestamp when window resets

    Tiers: free (100/min), pro (1,000/min), enterprise (5,000/min).

    ## Response Format
    - Single objects: `{ "data": { ... } }`
    - Lists: `{ "items": [...], "total": N, "limit": N, "offset": N }`
    - Errors: `{ "error": "code", "message": "description" }`
  version: 0.1.0
  contact:
    name: Pulse Support
    url: https://beinfi.com
  license:
    name: MIT
servers:
  - url: https://api.beinfi.com/api/v1
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key starting with `sk_live_`
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c3d4e5f6-a7b8-9012-cdef-345678901234"
        url:
          type: string
          format: uri
          example: "https://example.com/webhook"
        events:
          type: array
          items:
            type: string
          example: ["payment.confirmed"]
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - events
        - isActive
        - createdAt
    WebhookCreated:
      allOf:
        - $ref: "#/components/schemas/Webhook"
        - type: object
          properties:
            secret:
              type: string
              description: "HMAC signing secret (64-char hex). Only returned at creation. Store securely."
              example: "a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2"
          required:
            - secret
    CreateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: HTTPS URL to receive webhook POST requests
          example: "https://example.com/webhook"
        events:
          type: array
          items:
            type: string
          description: Event types to subscribe to
          example: ["payment.confirmed"]
      required:
        - url
        - events
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: "unauthorized"
        message:
          type: string
          description: Human-readable error description
          example: "Invalid API key"
      required:
        - error
        - message
  headers:
    X-RateLimit-Limit:
      description: Max requests per minute for your tier
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: Requests remaining in current window
      schema:
        type: integer
        example: 95
    X-RateLimit-Reset:
      description: Unix timestamp when rate limit window resets
      schema:
        type: integer
        example: 1705312260
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "unauthorized"
            message: "Invalid API key"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "rate_limit_exceeded"
            message: "Rate limit exceeded"
...
paths:
...
  /webhooks:
    post:
      operationId: createWebhook
      summary: Create a webhook subscription
      description: |
        Create a new webhook subscription. The response includes a `secret`
        (64-char hex string) used to verify webhook signatures. **This secret is
        only returned once** — store it securely.
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "200":
          description: Webhook created with signing secret
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/WebhookCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

    get:
      operationId: listWebhooks
      summary: List webhook subscriptions
      description: List all webhook subscriptions for the authenticated user.
      tags: [Webhooks]
      responses:
        "200":
          description: List of webhooks
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Webhook"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...